handle it gracefully if the input flake does not have a nixpkgs input
Admittedly it's not super common, but there are flakes that do not have
a nixpkgs input. I am using one that gets nixpkgs in a different way.

In the two cases where guestSystem != hostSystem, and where the
configuration is supplied as a module, there is no getting around *some*
dependence on nixpkgs, but we can fall back to <nixpkgs>.

In the case where guestSystem == hostSystem and the configuration
is already evaluated, we can do away with our dependence on nixpkgs
entirely. attrByPath is only needed if there is a dynamically sized list
being used as an attribute path. We are only using fixed sized lists and
can therefore use attribute path access as normal.